\name{scBIRD-package}
\alias{scBIRD-package}
\alias{scBIRD}
\docType{package}
\title{\packageTitle{scBIRD}}
\description{\packageDescription{scBIRD}}
\section{Package Content}{\packageIndices{scBIRD}}
\author{\packageAuthor{scBIRD}}
\section{Maintainer}{\packageMaintainer{scBIRD}}
\keyword{package}
